[app] Implementing history table.
I've created a table beforehand, so in this process we are going to do 2 different processes.
First, the mock script was adjusted to create a dataset with most recent records for each user. Then, we are going to save it to a user_score auto generated table (overwrite always).
Second, the mock script will also generate a history dataset containing all user history. Then, we are going to save it to a user_score_history which is a table created manually.